Tue, Feb 20, 2024, 11:05 AM +05:30

+++ main.cpp
@@ -1,16 +1,18 @@
 #include <iostream>
 
 using namespace std;
-int main() {
-    int a,b;
+
+int main()
+{
+    int a, b;
     float average;
 
     cin >> a;
     cin >> b;
 
-    cout << (float)a + (float)b << endl;
+    cout << ((float) a + b) << endl; // Cast a to float before addition
 
-    average = ((float)a + b)/2;
+    average = ((float) a + b) / 2;
 
     cout << "Hello, World! " << average << endl;
